pyproject.toml
[project]
name = "yt-transcriber"
version = "0.1.0"
description = "Download and transcribe YT videos"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "moviepy>=2.1.2",
    "openai-whisper>=20240930",
    "pytube>=15.0.0",
    "pytubefix>=8.13.1",
    "rich>=14.0.0",
]
